backend/services/learner-companion/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
sse-starlette==2.0.0
python-dotenv==1.0.1
openai==1.14.2
redis[asyncio]==5.0.1
httpx==0.27.0
psycopg2-binary
pydantic>=2.8.2
